refactor(flags): move definition of common flags to internal
