+++ cmd/ical/main.go
@@ -24,7 +24,7 @@ var (
 
 // init initializes the package by setting up flag options, log flags, and prefix.
 func init() {
-	obsidianutils.AddCommonFlagPrefixes()
+	internal.AddCommonFlagPrefixes()
 	flag.SetEnvPrefix("OBS_UTIL_ICAL")
 	flag.SetEnvPrefixForFlag("meeting-folder", "OBS_UTIL_AM")
 	flag.StringVar(&logLevel, "log-level", "info", "pass log level (debug/info/warn/error)")
