+++ cmd/obs-fm/main.go
@@ -22,7 +22,7 @@ var (
 
 // init initializes the package by setting up flag options, log flags, and prefix.
 func init() {
-	obsidianutils.AddCommonFlagPrefixes()
+	internal.AddCommonFlagPrefixes()
 	flag.SetEnvPrefix("OBS_UTIL_FM")
 	flag.StringVar(&logLevel, "log-level", "info", "pass log level (debug/info/warn/error)")
 	flag.StringVar(&dailyFolder, "daily-folder", "", "where to store the daily note inside the vault")
